feat[killall.py]Script to kill liveleech
FreeBSD jails are dumb and don't have proper service management. FreeBSD shell is not fully featured even when switching to bash. This kill-all script will look for all pid files in a particular directory and kill them. Useful if you need to restart all of your liveleech processes for a code change.